requirements.txt
numpy
scipy
psutil
shapely
pyntcore
matplotlib
pybresenham